// src/api/generated/schemas/duckResponse.ts
/**
 * Generated by orval v7.16.1 🍺
 * Do not edit manually.
 * Duck Party API
 * API for Duck Party backend service
 * OpenAPI spec version: 1.0
 */
import type { DuckAppearance } from "./duckAppearance";
import type { DuckUserResponse } from "./duckUserResponse";

export interface DuckResponse {
  appearance?: DuckAppearance;
  created_at?: string;
  dislikes_count?: number;
  id?: number;
  image?: string;
  likes_count?: number;
  name?: string;
  owner?: DuckUserResponse;
  owner_id?: number;
  rank?: number;
  updated_at?: string;
  x?: number;
  y?: number;
}